changed classes to fulfill special requirement
