+++ DonjinKrawler-Common/src/main/java/krawlercommon/enemies/Enemy.java
@@ -16,7 +16,7 @@ public abstract class Enemy extends EnemyTemplate implements Observer, KrawlerCl
     Random random = new Random();
     public transient Map<Phases, EnemyStrategy> strategies;
     transient EnemyStrategy currentStrategy;
-    int updateIntervalSeconds;
+    public int updateIntervalSeconds;
     int tick = 0;
     String info;
     private String name;
